[LC-994] Re-locate initialize lft logic into ConsensusRunner
